requirements.txt
flask 
flask-cors 
python-dotenv
sqlalchemy 
psycopg2-binary 
mysql-connector-python
PyJWT